Vertex: Move vertex attribute decl in vertex
